Introduce DatabaseRegionConvertible
